+++ packages/yourloops/test/components/menus/team-menu.test.tsx
@@ -29,53 +29,55 @@ import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
 import { act } from "react-dom/test-utils";
 import { Router } from "react-router-dom";
-import { loadTeams, Team, TeamContextProvider } from "../../../lib/team";
-import TeamMenu from "../../../components/menus/team-menu";
-import { teamAPI } from "../../lib/team/utils";
-import { AuthContextProvider, Session } from "../../../lib/auth";
-import { createAuthHookStubs } from "../../lib/auth/utils";
-import { loggedInUsers } from "../../common";
-import { NotificationContextProvider } from "../../../lib/notifications";
-import { stubNotificationContextValue } from "../../lib/notifications/utils";
-import { triggerMouseEvent } from "../../common/utils";
-import { createMemoryHistory } from "history";
-import * as shareLib from "../../../lib/share";
 
-jest.mock("../../../lib/share");
+import * as teamHookMock from "../../../lib/team";
+import { Team } from "../../../lib/team";
+import TeamMenu from "../../../components/menus/team-menu";
+import { buildTeam, triggerMouseEvent } from "../../common/utils";
+import { createMemoryHistory } from "history";
+import * as authHookMock from "../../../lib/auth";
+import { Session } from "../../../lib/auth";
+import User from "../../../lib/auth/user";
+import DirectShareApi from "../../../lib/share/direct-share-api";
+
+jest.mock("../../../lib/team");
+jest.mock("../../../lib/auth");
 describe("Team Menu", () => {
   let container: HTMLElement | null = null;
-  let filteredTeams: Team[];
-  const { hcpSession } = loggedInUsers;
   const history = createMemoryHistory({ initialEntries: ["/"] });
+  const teams: Team[] = [buildTeam("team1Id", []), buildTeam("team1Id", [])];
+  const session: Session = { user: {} as User, sessionToken: "fakeSessionToken", traceToken: "fakeTraceToken" };
 
   function openMenu(): void {
     const teamMenu = document.getElementById("team-menu");
     triggerMouseEvent("click", teamMenu);
   }
 
-  async function mountComponent(session: Session): Promise<void> {
-    const context = createAuthHookStubs(session);
-    const { teams } = await loadTeams(hcpSession, teamAPI.fetchTeams, teamAPI.fetchPatients);
-    filteredTeams = teams.filter(team => team.code !== "private");
-
-    await act(() => {
-      return new Promise((resolve) => {
-        render(
-          <Router history={history}>
-            <AuthContextProvider value={context}>
-              <NotificationContextProvider value={stubNotificationContextValue}>
-                <TeamContextProvider teamAPI={teamAPI}>
-                  <TeamMenu />
-                </TeamContextProvider>
-              </NotificationContextProvider>
-            </AuthContextProvider>
-          </Router>, container, resolve);
-      });
+  function mountComponent(): void {
+    act(() => {
+      render(
+        <Router history={history}>
+          <TeamMenu />
+        </Router>, container);
     });
   }
 
   beforeAll(() => {
-    jest.spyOn(shareLib, "getDirectShares").mockResolvedValue([]);
+    jest.spyOn(DirectShareApi, "getDirectShares").mockResolvedValue([]);
+    (teamHookMock.TeamContextProvider as jest.Mock) = jest.fn().mockImplementation(({ children }) => {
+      return children;
+    });
+    (teamHookMock.useTeam as jest.Mock).mockImplementation(() => {
+      return { teams };
+    });
+    (authHookMock.AuthContextProvider as jest.Mock) = jest.fn().mockImplementation(({ children }) => {
+      return children;
+    });
+    (authHookMock.useAuth as jest.Mock).mockImplementation(() => {
+      return {
+        session: () => session,
+      };
+    });
   });
 
   beforeEach(() => {
@@ -91,22 +93,22 @@ describe("Team Menu", () => {
     }
   });
 
-  it("should display number of teams user belongs to", async () => {
-    await mountComponent(hcpSession);
+  it("should display number of teams user belongs to", () => {
+    mountComponent();
     const teamBadge = container.querySelector("#team-menu-count-badge");
-    expect(filteredTeams.length.toString()).toEqual(teamBadge.textContent);
+    expect(teams.length.toString()).toEqual(teamBadge.textContent);
   });
 
-  it("should list all the teams user belongs to", async () => {
-    await mountComponent(hcpSession);
+  it("should list all the teams user belongs to", () => {
+    mountComponent();
     openMenu();
     const teamListItems = document.querySelectorAll("div.team-menu-list-item");
-    expect(filteredTeams.length).toEqual(teamListItems.length);
+    expect(teams.length).toEqual(teamListItems.length);
   });
 
-  it("should redirect to team details page when clicking on a team name", async () => {
-    await mountComponent(hcpSession);
-    const teamToSelect = filteredTeams[0];
+  it("should redirect to team details page when clicking on a team name", () => {
+    mountComponent();
+    const teamToSelect = teams[0];
     openMenu();
     const teamElement = document.getElementById(`team-menu-list-item-${teamToSelect.id}`);
     triggerMouseEvent("click", teamElement);