+++ packages/yourloops/test/pages/stepper/signup-profile-form.test.tsx
@@ -31,47 +31,41 @@ import { render, unmountComponentAtNode } from "react-dom";
 import { act } from "@testing-library/react-hooks/dom";
 import _ from "lodash";
 
-import { SignUpFormStateProvider, useSignUpFormState } from "../../../pages/signup/signup-formstate-context";
+import { SignUpFormStateProvider } from "../../../pages/signup/signup-formstate-context";
 import SignUpProfileForm from "../../../pages/signup/signup-profile-form";
+import * as authHookMock from "../../../lib/auth";
+import User from "../../../lib/auth/user";
+import { UserRoles } from "../../../models/shoreline";
 
-
-function FakeHcpSelector(): JSX.Element {
-  const { state, dispatch } = useSignUpFormState();
-
-  React.useEffect(() => {
-    if (state.formValues.accountRole !== "hcp") {
-      dispatch({
-        type: "EDIT_FORMVALUE",
-        key: "accountRole",
-        value: "hcp",
-      });
-    }
-  }, [dispatch, state.formValues.accountRole]);
-
-  return null;
-}
-
+jest.mock("../../../lib/auth");
 describe("Signup profile form", () => {
   let container: HTMLElement | null = null;
 
-  const mountComponent = async (hcp: boolean): Promise<void> => {
-    await act(() => {
-      return new Promise((resolve) => {
-
-        render(
-          <SignUpFormStateProvider>
-            {hcp &&
-              <FakeHcpSelector />
-            }
-            <SignUpProfileForm handleBack={_.noop} handleNext={_.noop} />
-          </SignUpFormStateProvider>, container, resolve);
-      });
+  const mountComponent = (): void => {
+    act(() => {
+      render(
+        <SignUpFormStateProvider>
+          <SignUpProfileForm handleBack={_.noop} handleNext={_.noop} />
+        </SignUpFormStateProvider>, container);
     });
   };
+
+  beforeAll(() => {
+    (authHookMock.AuthContextProvider as jest.Mock) = jest.fn().mockImplementation(({ children }) => {
+      return children;
+    });
+  });
 
   beforeEach(() => {
     container = document.createElement("div");
     document.body.appendChild(container);
+    (authHookMock.useAuth as jest.Mock).mockImplementation(() => {
+      return {
+        user: {
+          role: UserRoles.hcp,
+        } as User,
+      };
+    });
   });
 
   afterEach(() => {
@@ -82,14 +76,21 @@ describe("Signup profile form", () => {
     }
   });
 
-  it("should not render the drop down list when caregiver", async () => {
-    await mountComponent(false);
+  it("should not render the drop down list when caregiver", () => {
+    (authHookMock.useAuth as jest.Mock).mockImplementation(() => {
+      return {
+        user: {
+          role: UserRoles.caregiver,
+        } as User,
+      };
+    });
+    mountComponent();
     const dropDownList = document.querySelector("#hcp-profession-selector");
     expect(dropDownList).toBeNull();
   });
 
-  it("should render the drop down list when HCP", async () => {
-    await mountComponent(true);
+  it("should render the drop down list when HCP", () => {
+    mountComponent();
     const dropDownList = document.querySelector("#hcp-profession-selector");
     expect(dropDownList).not.toBeNull();
   });